add missing NCL screens
