+++ apps/native-component-list/src/navigation/ExpoComponentsStackNavigator.tsx
@@ -13,6 +13,18 @@ const Stack = createStackNavigator();
 export const Screens = [
   {
     getComponent() {
+      return optionalRequire(() => require('../screens/FacebookAdsScreen'));
+    },
+    name: 'FacebookAds',
+  },
+  {
+    getComponent() {
+      return optionalRequire(() => require('../screens/DrawerLayoutAndroidScreen'));
+    },
+    name: 'DrawerLayoutAndroid',
+  },
+  {
+    getComponent() {
       return optionalRequire(() => require('../screens/AdMobScreen'));
     },
     name: 'Admob',
